喵

+++ src/io/kurumi/ntt/fragment/group/JoinCaptcha.java
@@ -1,38 +1,33 @@
 package io.kurumi.ntt.fragment.group;
 
 import cn.hutool.captcha.generator.RandomGenerator;
-import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.NumberUtil;
 import cn.hutool.core.util.RandomUtil;
 import com.pengrad.telegrambot.model.User;
+import com.pengrad.telegrambot.model.request.ParseMode;
+import com.pengrad.telegrambot.request.DeleteMessage;
+import com.pengrad.telegrambot.request.SendPhoto;
+import com.pengrad.telegrambot.response.SendResponse;
 import io.kurumi.ntt.db.GroupData;
+import io.kurumi.ntt.db.PointData;
 import io.kurumi.ntt.db.UserData;
 import io.kurumi.ntt.fragment.BotFragment;
 import io.kurumi.ntt.fragment.Fragment;
+import io.kurumi.ntt.model.Callback;
 import io.kurumi.ntt.model.Msg;
 import io.kurumi.ntt.model.request.ButtonMarkup;
+import io.kurumi.ntt.model.request.Send;
 import io.kurumi.ntt.utils.Img;
+import io.kurumi.ntt.utils.NTT;
 import java.awt.Color;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Collection;
-import java.util.Collections;
-import io.kurumi.ntt.db.PointData;
-import com.pengrad.telegrambot.request.SendPhoto;
-import com.pengrad.telegrambot.model.request.ParseMode;
-import com.pengrad.telegrambot.response.SendResponse;
-import io.kurumi.ntt.fragment.group.JoinCaptcha.AuthCache;
-import java.util.TimerTask;
-import io.kurumi.ntt.model.Callback;
-import io.kurumi.ntt.utils.NTT;
-import java.util.Date;
-import io.kurumi.ntt.fragment.group.JoinCaptcha.VerifyCode;
-import com.pengrad.telegrambot.request.DeleteMessage;
-import org.omg.CORBA.CustomMarshal;
 import java.util.Map;
-import io.kurumi.ntt.model.request.Send;
+import java.util.TimerTask;
 
 public class JoinCaptcha extends Fragment {
 
@@ -886,6 +881,15 @@ public class JoinCaptcha extends Fragment {
 
 				} else if (POINT_ANSWER.equals(point)) {
 
+						if (!user.id.equals(target)) {
+
+								callback.alert("这个验证不针对你。");
+
+								return;
+
+						}
+						
+						
 						if (auth == null) {
 
 								callback.delete();
